requirements.txt
.
django-bootstrap-v5